Catching exceptions to keep running bot

+++ src/main/java/dto/Chat.java
@@ -1,4 +1,4 @@
 package dto;
 
-public record Chat(int id) {
+public record Chat(long id) {
 }
